refactor: create useEditProfileForm (#46)
